+++ test/fixtures/rspec_style_app/spec/widget_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.describe Widget do
+  it "has a nil name by default" do
+    expect(Widget.new.name).to be_nil
+  end
+end